src/python/requirements.txt
PyMuPDF
pytesseract
pillow
pandas
numpy
scikit-learn
spacy
textblob
beautifulsoup4
requests
flask
python-docx
matplotlib
reportlab
werkzeug
pyod
google-genai
python-dotenv
duckduckgo-search
faiss-cpu
camelot-py[cv]
opencv-python-headless
tabula-py
feedparser
sentence-transformers
yfinance
